requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
hashin==0.15.0
idna==2.9
packaging==20.4
pip-api==0.0.14
psycopg2==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
